[Refactor] move "non-service" code to `notificationInstance`
